requirements.txt
sqlalchemy>=2.0
psycopg2-binary
alembic
pydantic>=2.0
python-dotenv
pandas
numpy
scipy
scikit-learn
torch
mlflow
jupyter
matplotlib
seaborn
requests
holidays
